refactor: split panels.tsx into DataSourcePanel + JobHistoryPanel
Extract DataSourceSection into DataSourcePanel.tsx and
JobHistorySection + StatusBadge into JobHistoryPanel.tsx.
panels.tsx becomes a temporary barrel (removal target: 2026-03-14).
